+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
  - Fix issue with SHT31 Smart Gadget disconnect error-handling
  - Prevent dashboard camera streaming if using the fswebcam library ([#630](https://github.com/kizniche/mycodo/issues/630))
+ - Fix number of line characters for 20x4 LCDs ([#627](https://github.com/kizniche/mycodo/issues/627))
 
 
 ## 7.2.2 (2019-02-08)